requirements.txt
pygame-ce
discord-rpc